refactor: constants.ts를 활용하여 상수를 import 하도록 변경 (#47)
