+++ src/pages/user/Schedule/ScheduleHeader/MonthPickerDetail/index.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { colors, border } from '../../../../../styles';
-import { MONTHS } from '../constants';
+import { MONTHS } from '../../constants';
 
 interface MonthPickerDetailProps {
   currentMonth: number;
