+++ src/pages/user/Schedule/ScheduleSideBar/index.tsx
@@ -1,37 +1,7 @@
 import styled from 'styled-components';
 import TeamList from '../core/TeamList';
 import { border } from '../../../../styles';
-
-const teams = [
-  {
-    name: '응급 1팀',
-    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
-  },
-  {
-    name: '응급 2팀',
-    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
-  },
-  {
-    name: '응급 3팀',
-    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
-  },
-  {
-    name: '응급 4팀',
-    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
-  },
-  {
-    name: '응급 5팀',
-    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
-  },
-  {
-    name: '응급 6팀',
-    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
-  },
-  {
-    name: '응급 7팀',
-    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
-  },
-];
+import { TEAMS } from '../constants';
 
 interface ScheduleSideBarProps {
   isSixWeek: boolean;
@@ -41,7 +11,7 @@ const ScheduleSideBar = ({ isSixWeek }: ScheduleSideBarProps) => {
   return (
     <S.Wrapper isSixWeek={isSixWeek}>
       <S.Content>
-        {teams.map((team) => (
+        {TEAMS.map((team) => (
           <TeamList key={team.name} name={team.name} members={team.members} />
         ))}
       </S.Content>
