+++ src/pages/user/Schedule/constants.ts
@@ -1,0 +1,60 @@
+export const MONTHS: Record<number, string> = {
+  1: 'Jan',
+  2: 'Feb',
+  3: 'Mar',
+  4: 'Apr',
+  5: 'May',
+  6: 'Jun',
+  7: 'Jul',
+  8: 'Aug',
+  9: 'Sep',
+  10: 'Oct',
+  11: 'Nov',
+  12: 'Dec',
+};
+
+export const DAYS_OF_WEEK_LIST: string[] = [
+  'Mon',
+  'Tue',
+  'Wed',
+  'Thu',
+  'Fri',
+  'Sat',
+  'Sun',
+];
+
+export const HOURS: number[] = [
+  0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
+  22, 23, 24,
+];
+
+export const TEAMS = [
+  {
+    name: '응급 1팀',
+    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
+  },
+  {
+    name: '응급 2팀',
+    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
+  },
+  {
+    name: '응급 3팀',
+    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
+  },
+  {
+    name: '응급 4팀',
+    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
+  },
+  {
+    name: '응급 5팀',
+    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
+  },
+  {
+    name: '응급 6팀',
+    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
+  },
+  {
+    name: '응급 7팀',
+    members: ['김하나', '김둘둘', '김셋셋', '김넷넷', '김다섯', '김여섯'],
+  },
+];
